moved solver added Doc
